refactor: improve graph editing
